ml/ml/api_requirements.txt
tensorflow==2.13.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.13.0rc1
matplotlib
numpy
opencv-python
imutils
scipy